refractor: FE OverviewPage -> Move modifyStudentHandler to DiplomaCardContainer
